Re-organise ImageViewController
