+++ piwigo/Image/Extensions/ImageViewController+Buttons.swift
@@ -1,0 +1,14 @@
+//
+//  ImageViewController+Buttons.swift
+//  piwigo
+//
+//  Created by Eddy Lelièvre-Berna on 19/06/2022.
+//  Copyright © 2022 Piwigo.org. All rights reserved.
+//
+
+import Foundation
+
+extension ImageViewController
+{
+    
+}
