+++ piwigo/Image/ImageDescriptionView.swift
